Leaderboard System
Fully functional leaderboard system:
When player reach end of level, he can continue to the next level, go
to leaderboards or main menu.
Leaderboards save the total of berries picked, the time that i took to
complete the level and the combined total of points. Additionally the
user can input his/her own name to save on the leaderboards